Timestamp and epochnanos variables must use Int64 type, or wont work in 32 bits builds

+++ Sources/OpenTelemetrySdk/Trace/ReadableSpan.swift
@@ -32,5 +32,5 @@ public protocol ReadableSpan: Span {
     var hasEnded: Bool { get }
 
     /// Returns the latecy of the {@code Span} in nanos. If still active then returns now() - start time.
-    var latencyNanos: Int { get }
+    var latencyNanos: Int64 { get }
 }
